backend/requirements.txt
#python version 3.10
asgiref==3.7.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.3.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-import-export==3.3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
fpdf==1.7.2
idna==3.6
import-export==0.3.1
lxml==5.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
plum-dispatch==1.7.4
PyJWT==2.8.0
PyMuPDF==1.23.19
PyMuPDFb==1.23.9
PyPDF2==3.0.1
python-docx==1.1.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
